0.2.0.
1. New feature – Inverse order
Now you can show a playlist in inverse order.

2. New feature – Error Message
When an error occurred, the module show an error message with a error
code.

3. List error corrected
There is an error showing NULL items if the number of item in the list
is less than the designated number of list in configuration.

4. Typing error corrected
The errors such as 유투브(유튜브) and so on are corrected.

5. SNS icon corrected
The default skin includes XEIcon now.